// markets/wsteth-arbitrum-jun2024/reservesConfig.ts
import { eContractid, IReserveParams } from "../../helpers/types";

import {
  rateStrategyVolatileOne,
  rateStrategyStableOne,
  rateStrategyStableTwo,
} from "./rateStrategies";

export const strategyWstETH: IReserveParams = {
  strategy: rateStrategyStableOne,
  baseLTVAsCollateral: "9500",
  liquidationThreshold: "9800",
  liquidationBonus: "10400",
  liquidationProtocolFee: "1000",
  borrowingEnabled: true,
  stableBorrowRateEnabled: true,
  flashLoanEnabled: true,
  reserveDecimals: "18",
  aTokenImpl: eContractid.AToken,
  reserveFactor: "1000",
  supplyCap: "0",
  borrowCap: "0",
  debtCeiling: "0",
  borrowableIsolation: false,
};

export const strategyPtWstETH: IReserveParams = {
  strategy: rateStrategyStableOne,
  baseLTVAsCollateral: "9500",
  liquidationThreshold: "9800",
  liquidationBonus: "10400",
  liquidationProtocolFee: "1000",
  borrowingEnabled: true,
  stableBorrowRateEnabled: true,
  flashLoanEnabled: true,
  reserveDecimals: "18",
  aTokenImpl: eContractid.AToken,
  reserveFactor: "1000",
  supplyCap: "0",
  borrowCap: "0",
  debtCeiling: "0",
  borrowableIsolation: false,
};
